tests/mediacapture-depth/BAK/GUM-impossible-constraint-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Trivial mandatory constraint in getUserMedia</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#constrainable-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<h2>Description</h2>
<p>
  This test checks that setting an impossible mandatory
constraint (width &gt;=1G) in getUserMedia works.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    Use a test device with depth camera(embedded or external).
  </li>
  <li>
    When prompted, accept to share your depth stream.
  </li>
</ol>

<div id="log"></div>

<script>

promise_test(t => {
  // Note - integer conversion is weird for +inf and numbers > 2^32, so we
  // use a number less than 2^32 for testing.
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}, width: {min: 100000000}}})
  .then(stream => {
    assert_unreached("a Video stream of width 100M cannot be created");
  })
  .catch(error => {
    assert_equals(error.name, "OverconstrainedError", "An impossible constraint triggers a OverconstrainedError");
    assert_equals(error.constraint, "width", "The name of the not satisfied error is given in error.constraint");
  });
}, "Tests that setting an impossible constraint in getUserMedia fails");

</script>
